ADD Doctrine DBAL foreign key migration feature
Adds a before and after task using a dedicated migrator interface,
a helper for users and a foreign key data object.
This feature creates temporary columns in both local and foreign
tables before migration. The filling of those columns is left to
the developer. Then the relations are corrected to their real
values and temporary columns are deleted.