REMOVE v0.x code
Removes the Fregata\Migrator namespace and the related tests, and
the old "kernel".
Removes PHP-DI dependency as it is now replaced by Symfony
DependencyInjection.
Changes the required version of symfony/* packages to allow v4.4+.
Removes unnecessary services in the development Docker Compose file
and the pdo_pgsql extension from the PHP service's Dockerfile.